use duration as my working time in dashboard widget
