+++ src/Widget/Type/DailyWorkingTimeChart.php
@@ -11,17 +11,21 @@ namespace App\Widget\Type;
 
 use App\Entity\Activity;
 use App\Entity\Project;
+use App\Entity\Timesheet;
+use App\Entity\User;
+use App\Model\Statistic\Day;
+use App\Repository\TimesheetRepository;
 use App\Timesheet\DateTimeFactory;
-use App\Widget\DataProvider\DailyWorkingTimeChartProvider;
 use App\Widget\WidgetInterface;
 use DateTime;
+use DateTimeInterface;
 
 /**
  * This is rendered inside the PaginatedWorkingTimeChart.
  */
 final class DailyWorkingTimeChart extends AbstractWidget
 {
-    public function __construct(private DailyWorkingTimeChartProvider $dailyWorkingTimeChartProvider)
+    public function __construct(private readonly TimesheetRepository $repository)
     {
     }
 
@@ -71,7 +75,7 @@ final class DailyWorkingTimeChart extends AbstractWidget
         $begin = $options['begin'];
         if (!($begin instanceof \DateTimeInterface)) {
             if (\is_string($begin)) {
-                $begin = new DateTime($begin, new \DateTimeZone($user->getTimezone()));
+                $begin = new \DateTimeImmutable($begin, new \DateTimeZone($user->getTimezone()));
             } else {
                 $begin = $dateTimeFactory->getStartOfWeek();
             }
@@ -80,14 +84,14 @@ final class DailyWorkingTimeChart extends AbstractWidget
         $end = $options['end'];
         if (!($end instanceof \DateTimeInterface)) {
             if (\is_string($end)) {
-                $end = new DateTime($end, new \DateTimeZone($user->getTimezone()));
+                $end = new \DateTimeImmutable($end, new \DateTimeZone($user->getTimezone()));
             } else {
                 $end = $dateTimeFactory->getEndOfWeek($begin);
             }
         }
 
         $activities = [];
-        $statistics = $this->dailyWorkingTimeChartProvider->getData($user, $begin, $end);
+        $statistics = $this->getPreparedData($user, $begin, $end);
 
         foreach ($statistics as $day) {
             foreach ($day->getDetails() as $entry) {
@@ -124,5 +128,127 @@ final class DailyWorkingTimeChart extends AbstractWidget
     public function getTemplateName(): string
     {
         return 'widget/widget-dailyworkingtimechart.html.twig';
+    }
+
+    /**
+     * @return list<array{'duration': int, 'billable': int, 'month': numeric-string, 'year': numeric-string, 'day': numeric-string, 'details': array<int|string, array<mixed>>}>
+     */
+    private function getDailyData(DateTimeInterface $begin, DateTimeInterface $end, ?User $user = null): array
+    {
+        $qb = $this->repository->createQueryBuilder('t');
+
+        $qb->select('t, p, a, c')
+            ->andWhere($qb->expr()->isNotNull('t.end'))
+            ->andWhere($qb->expr()->between('t.begin', ':begin', ':end'))
+            ->orderBy('t.begin', 'DESC')
+            ->setParameter('begin', $begin)
+            ->setParameter('end', $end)
+            ->leftJoin('t.activity', 'a')
+            ->leftJoin('t.project', 'p')
+            ->leftJoin('p.customer', 'c')
+        ;
+
+        if (null !== $user) {
+            $qb
+                ->andWhere($qb->expr()->eq('t.user', ':user'))
+                ->setParameter('user', $user)
+            ;
+        }
+
+        $timesheets = $qb->getQuery()->getResult();
+
+        $results = [];
+        /** @var Timesheet $result */
+        foreach ($timesheets as $result) {
+            /** @var DateTime $beginTmp */
+            $beginTmp = $result->getBegin();
+
+            $dateKey = $beginTmp->format('Ymd');
+
+            if (!isset($results[$dateKey])) {
+                $results[$dateKey] = [
+                    'duration' => 0,
+                    'billable' => 0, // duration
+                    'month' => $beginTmp->format('n'),
+                    'year' => $beginTmp->format('Y'),
+                    'day' => $beginTmp->format('j'),
+                    'details' => []
+                ];
+            }
+            $duration = $result->getDuration() ?? 0;
+
+            $results[$dateKey]['duration'] += $duration;
+            if ($result->isBillable()) {
+                $results[$dateKey]['billable'] += $duration;
+            }
+            $detailsId =
+                $result->getProject()->getCustomer()->getId()
+                . '_' . ($result->getProject()?->getId() ?? '')
+                . '_' . ($result->getActivity()?->getId() ?? '')
+            ;
+
+            if (!isset($results[$dateKey]['details'][$detailsId])) {
+                $results[$dateKey]['details'][$detailsId] = [
+                    'project' => $result->getProject(),
+                    'activity' => $result->getActivity(),
+                    'duration' => 0,
+                    'billable' => 0, // duration
+                ];
+            }
+
+            $results[$dateKey]['details'][$detailsId]['duration'] += $duration;
+            if ($result->isBillable()) {
+                $results[$dateKey]['details'][$detailsId]['billable'] += $duration;
+            }
+        }
+
+        ksort($results);
+
+        foreach ($results as $key => $value) {
+            $results[$key]['details'] = array_values($value['details']);
+        }
+
+        return array_values($results);
+    }
+
+    /**
+     * @return Day[]
+     * @throws \Exception
+     */
+    private function getPreparedData(User $user, DateTimeInterface $begin, DateTimeInterface $end): array
+    {
+        /** @var Day[] $days */
+        $days = [];
+
+        // prefill the array
+        $tmp = DateTime::createFromInterface($end);
+        $until = (int) $begin->format('Ymd');
+        while ((int) $tmp->format('Ymd') >= $until) {
+            $last = clone $tmp;
+            $days[$last->format('Ymd')] = new Day($last, 0, 0.00);
+            $tmp->modify('-1 day');
+        }
+
+        // TODO replace with TimesheetStatisticService::getDailyStatistics()
+        $results = $this->getDailyData($begin, $end, $user);
+
+        foreach ($results as $statRow) {
+            $dateTime = DateTime::createFromInterface($begin);
+            $dateTime->setDate((int) $statRow['year'], (int) $statRow['month'], (int) $statRow['day']);
+            $dateTime->setTime(0, 0, 0);
+            $day = new Day($dateTime, (int) $statRow['duration'], 0.00); // rate is not used in frontend
+            $day->setTotalDurationBillable($statRow['billable']);
+            $day->setDetails($statRow['details']);
+            $dateKey = $dateTime->format('Ymd');
+            // make sure entries from other timezones are filtered
+            if (!\array_key_exists($dateKey, $days)) {
+                continue;
+            }
+            $days[$dateKey] = $day;
+        }
+
+        ksort($days);
+
+        return array_values($days);
     }
 }
